fix: propagate MsEntropyCleanSpectrum generic config

+++ tests/test_ms_entropy_validation.rs
@@ -44,7 +44,7 @@ fn validate_linear_entropy_against_ms_entropy() {
         };
 
         let cleaner = MsEntropyCleanSpectrum::<f32>::builder()
-            .min_ms2_difference_in_da(2.0 * f64::from(tolerance))
+            .min_ms2_difference_in_da(2.0 * tolerance)
             .build()
             .expect("valid ms_entropy-compatible cleaner config");
         let left = cleaner.process(&left);
